Add video player composable
