+++ posts/src/main/java/com/neaniesoft/vermilion/posts/data/PostRecordConverters.kt
@@ -27,10 +27,10 @@ import com.neaniesoft.vermilion.posts.domain.entities.Score
 import com.neaniesoft.vermilion.posts.domain.entities.TextPostSummary
 import com.neaniesoft.vermilion.posts.domain.entities.ThumbnailSummary
 import com.neaniesoft.vermilion.posts.domain.entities.UriImage
-import com.neaniesoft.vermilion.posts.domain.entities.VideoDescriptor
-import com.neaniesoft.vermilion.posts.domain.entities.VideoHeight
+import com.neaniesoft.vermilion.ui.videos.VideoDescriptor
+import com.neaniesoft.vermilion.ui.videos.VideoHeight
 import com.neaniesoft.vermilion.posts.domain.entities.VideoPostSummary
-import com.neaniesoft.vermilion.posts.domain.entities.VideoWidth
+import com.neaniesoft.vermilion.ui.videos.VideoWidth
 import org.commonmark.node.Document
 import org.commonmark.parser.Parser
 import java.time.Clock
