+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/VideoDescriptor.kt
@@ -1,0 +1,11 @@
+package com.neaniesoft.vermilion.ui.videos
+
+import android.net.Uri
+
+data class VideoDescriptor(
+    val width: VideoWidth,
+    val height: VideoHeight,
+    val dash: Uri,
+    val hls: Uri,
+    val fallback: Uri
+)
