+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/VideoHeight.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.ui.videos
+
+@JvmInline
+value class VideoHeight(val value: Int)
