+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/VideoWidth.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.ui.videos
+
+@JvmInline
+value class VideoWidth(val value: Int)
